requirement.txt
python-dotenv
pandas 
pyodbc
pymilvus
langchain-google-genai
google-generativeai
google-ai-generativelanguage==0.6.18
ipywidgets
ipykernel
streamlit
sentence_transformers
numpy
fastapi
uvicorn